backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==1.4.46
pydantic==1.10.7
scikit-learn==1.3.0
numpy==1.24.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
